+++ src/pages/about.tsx
@@ -1,11 +1,11 @@
 import App from '../components/App'
-import Sample from '../components/Sample'
+import ProblemItemList from '../components/ProblemItemList'
 
 export default function About() {
   return (
     <App>
       <p>About Page</p>
-      <Sample></Sample>
+      <ProblemItemList></ProblemItemList>
     </App>
   )
 }
